refactor: routes groups 모듈화
Related to: #143
